backend/requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
coreschema==0.0.4
Django==3.0.4
djangorestframework==3.11.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
